* Fixed : session management for logged in users
* AdminInitializer to premake an admin account for testing purposes

* Enable access to H2-Console with dev property

* Exception handling in SecurityConfig (WIP)

* Give access for login/register pages to only guest (not logged in) users

+++ cydomotix/src/main/java/com/example/cydomotix/Repository/UserRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     // Here, include custom CRUD methods
